pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.72.0",
    "email-validator>=2.3.0",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.121.0",
    "langchain>=1.0.4",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.2",
    "openai>=2.7.1",
    "passlib[bcrypt]>=1.7.4",
    "pdf2image>=1.17.0",
    "pillow>=12.0.0",
    "pydantic-settings>=2.11.0",
    "pytesseract>=0.3.13",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.27",
    "uvicorn[standard]>=0.38.0",
]
